Struct aws_sdk_lexmodelbuilding::client::fluent_builders::PutBot

pub struct PutBot<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> PutBot<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<PutBotOutput, SdkError<PutBotError>> where
R::Policy: SmithyRetryPolicy<PutBotInputOperationOutputAlias, PutBotOutput, PutBotError, PutBotInputOperationRetryAlias>,

pub fn name(self, inp: impl Into<String>) -> Self

The name of the bot. The name is not case sensitive.

pub fn set_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the bot.

pub fn set_description(self, input: Option<String>) -> Self

pub fn intents(self, inp: impl Into<Intent>) -> Self

Appends an item to intents.

To override the contents of this collection use set_intents.

An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see how-it-works.

pub fn set_intents(self, input: Option<Vec<Intent>>) -> Self

pub fn enable_model_improvements(self, inp: bool) -> Self

Set to true to enable access to natural language understanding improvements.

When you set the enableModelImprovements parameter to true you can use the nluIntentConfidenceThreshold parameter to configure confidence scores. For more information, see Confidence Scores.

You can only set the enableModelImprovements parameter in certain Regions. If you set the parameter to true, your bot has access to accuracy improvements.

The Regions where you can set the enableModelImprovements parameter to true are:

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default. In these Regions setting the parameter to false throws a ValidationException exception.

pub fn set_enable_model_improvements(self, input: Option<bool>) -> Self

pub fn nlu_intent_confidence_threshold(self, inp: f64) -> Self

Determines the threshold where Amazon Lex will insert the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent and AMAZON.KendraSearchIntent are only inserted if they are configured for the bot.

You must set the enableModelImprovements parameter to true to use confidence scores in the following regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default.

For example, suppose a bot is configured with the confidence threshold of 0.80 and the AMAZON.FallbackIntent. Amazon Lex returns three alternative intents with the following confidence scores: IntentA (0.70), IntentB (0.60), IntentC (0.50). The response from the PostText operation would be:

AMAZON.FallbackIntent

IntentA

IntentB

IntentC

pub fn set_nlu_intent_confidence_threshold(self, input: Option<f64>) -> Self

pub fn clarification_prompt(self, inp: Prompt) -> Self

When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeat the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field.

When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

If you have defined a fallback intent, it will be invoked if the clarification prompt is repeated the number of times defined in the maxAttempts field. For more information, see AMAZON.FallbackIntent.

If you don't define a clarification prompt, at runtime Amazon Lex will return a 400 Bad Request exception in three cases:

Follow-up prompt - When the user responds to a follow-up prompt but does not provide an intent. For example, in response to a follow-up prompt that says "Would you like anything else today?" the user says "Yes." Amazon Lex will return a 400 Bad Request exception because it does not have a clarification prompt to send to the user to get an intent.

Lambda function - When using a Lambda function, you return an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

PutSession operation - When using the PutSession operation, you send an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

pub fn set_clarification_prompt(self, input: Option<Prompt>) -> Self

pub fn abort_statement(self, inp: Statement) -> Self

When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then cancels the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type.

For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times.

For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

If you have defined a fallback intent the cancel statement will not be sent to the user, the fallback intent is used instead. For more information, see AMAZON.FallbackIntent.

pub fn set_abort_statement(self, input: Option<Statement>) -> Self

pub fn idle_session_ttl_in_seconds(self, inp: i32) -> Self

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation.

A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout.

For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over.

If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot.

The default is 300 seconds (5 minutes).

pub fn set_idle_session_ttl_in_seconds(self, input: Option<i32>) -> Self

pub fn voice_id(self, inp: impl Into<String>) -> Self

The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Voices in Amazon Polly in the Amazon Polly Developer Guide.

pub fn set_voice_id(self, input: Option<String>) -> Self

pub fn checksum(self, inp: impl Into<String>) -> Self

Identifies a specific revision of the $LATEST version.

When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

pub fn set_checksum(self, input: Option<String>) -> Self

pub fn process_behavior(self, inp: ProcessBehavior) -> Self

If you set the processBehavior element to BUILD, Amazon Lex builds the bot so that it can be run. If you set the element to SAVE Amazon Lex saves the bot, but doesn't build it.

If you don't specify this value, the default value is BUILD.

pub fn set_process_behavior(self, input: Option<ProcessBehavior>) -> Self

pub fn locale(self, inp: Locale) -> Self

Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot.

The default is en-US.

pub fn set_locale(self, input: Option<Locale>) -> Self

pub fn child_directed(self, inp: bool) -> Self

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

pub fn set_child_directed(self, input: Option<bool>) -> Self

pub fn detect_sentiment(self, inp: bool) -> Self

When set to true user utterances are sent to Amazon Comprehend for sentiment analysis. If you don't specify detectSentiment, the default is false.

pub fn set_detect_sentiment(self, input: Option<bool>) -> Self

pub fn create_version(self, inp: bool) -> Self

When set to true a new numbered version of the bot is created. This is the same as calling the CreateBotVersion operation. If you don't specify createVersion, the default is false.

pub fn set_create_version(self, input: Option<bool>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

A list of tags to add to the bot. You can only add tags when you create a bot, you can't use the PutBot operation to update the tags on a bot. To update tags, use the TagResource operation.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self