Struct aws_sdk_lexmodelbuilding::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSDeepSenseModelBuildingService.

This client allows ergonomic access to a AWSDeepSenseModelBuildingService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_bot_version(&self) -> CreateBotVersion<C, M, R>

Constructs a fluent builder for the CreateBotVersion operation.

See CreateBotVersion for more information about the operation and its arguments.

pub fn create_intent_version(&self) -> CreateIntentVersion<C, M, R>

Constructs a fluent builder for the CreateIntentVersion operation.

See CreateIntentVersion for more information about the operation and its arguments.

pub fn create_slot_type_version(&self) -> CreateSlotTypeVersion<C, M, R>

Constructs a fluent builder for the CreateSlotTypeVersion operation.

See CreateSlotTypeVersion for more information about the operation and its arguments.

pub fn delete_bot(&self) -> DeleteBot<C, M, R>

Constructs a fluent builder for the DeleteBot operation.

See DeleteBot for more information about the operation and its arguments.

pub fn delete_bot_alias(&self) -> DeleteBotAlias<C, M, R>

Constructs a fluent builder for the DeleteBotAlias operation.

See DeleteBotAlias for more information about the operation and its arguments.

pub fn delete_bot_channel_association(
&self
) -> DeleteBotChannelAssociation<C, M, R>

Constructs a fluent builder for the DeleteBotChannelAssociation operation.

See DeleteBotChannelAssociation for more information about the operation and its arguments.

pub fn delete_bot_version(&self) -> DeleteBotVersion<C, M, R>

Constructs a fluent builder for the DeleteBotVersion operation.

See DeleteBotVersion for more information about the operation and its arguments.

pub fn delete_intent(&self) -> DeleteIntent<C, M, R>

Constructs a fluent builder for the DeleteIntent operation.

See DeleteIntent for more information about the operation and its arguments.

pub fn delete_intent_version(&self) -> DeleteIntentVersion<C, M, R>

Constructs a fluent builder for the DeleteIntentVersion operation.

See DeleteIntentVersion for more information about the operation and its arguments.

pub fn delete_slot_type(&self) -> DeleteSlotType<C, M, R>

Constructs a fluent builder for the DeleteSlotType operation.

See DeleteSlotType for more information about the operation and its arguments.

pub fn delete_slot_type_version(&self) -> DeleteSlotTypeVersion<C, M, R>

Constructs a fluent builder for the DeleteSlotTypeVersion operation.

See DeleteSlotTypeVersion for more information about the operation and its arguments.

pub fn delete_utterances(&self) -> DeleteUtterances<C, M, R>

Constructs a fluent builder for the DeleteUtterances operation.

See DeleteUtterances for more information about the operation and its arguments.

pub fn get_bot(&self) -> GetBot<C, M, R>

Constructs a fluent builder for the GetBot operation.

See GetBot for more information about the operation and its arguments.

pub fn get_bot_alias(&self) -> GetBotAlias<C, M, R>

Constructs a fluent builder for the GetBotAlias operation.

See GetBotAlias for more information about the operation and its arguments.

pub fn get_bot_aliases(&self) -> GetBotAliases<C, M, R>

Constructs a fluent builder for the GetBotAliases operation.

See GetBotAliases for more information about the operation and its arguments.

pub fn get_bot_channel_association(&self) -> GetBotChannelAssociation<C, M, R>

Constructs a fluent builder for the GetBotChannelAssociation operation.

See GetBotChannelAssociation for more information about the operation and its arguments.

pub fn get_bot_channel_associations(&self) -> GetBotChannelAssociations<C, M, R>

Constructs a fluent builder for the GetBotChannelAssociations operation.

See GetBotChannelAssociations for more information about the operation and its arguments.

pub fn get_bots(&self) -> GetBots<C, M, R>

Constructs a fluent builder for the GetBots operation.

See GetBots for more information about the operation and its arguments.

pub fn get_bot_versions(&self) -> GetBotVersions<C, M, R>

Constructs a fluent builder for the GetBotVersions operation.

See GetBotVersions for more information about the operation and its arguments.

pub fn get_builtin_intent(&self) -> GetBuiltinIntent<C, M, R>

Constructs a fluent builder for the GetBuiltinIntent operation.

See GetBuiltinIntent for more information about the operation and its arguments.

pub fn get_builtin_intents(&self) -> GetBuiltinIntents<C, M, R>

Constructs a fluent builder for the GetBuiltinIntents operation.

See GetBuiltinIntents for more information about the operation and its arguments.

pub fn get_builtin_slot_types(&self) -> GetBuiltinSlotTypes<C, M, R>

Constructs a fluent builder for the GetBuiltinSlotTypes operation.

See GetBuiltinSlotTypes for more information about the operation and its arguments.

pub fn get_export(&self) -> GetExport<C, M, R>

Constructs a fluent builder for the GetExport operation.

See GetExport for more information about the operation and its arguments.

pub fn get_import(&self) -> GetImport<C, M, R>

Constructs a fluent builder for the GetImport operation.

See GetImport for more information about the operation and its arguments.

pub fn get_intent(&self) -> GetIntent<C, M, R>

Constructs a fluent builder for the GetIntent operation.

See GetIntent for more information about the operation and its arguments.

pub fn get_intents(&self) -> GetIntents<C, M, R>

Constructs a fluent builder for the GetIntents operation.

See GetIntents for more information about the operation and its arguments.

pub fn get_intent_versions(&self) -> GetIntentVersions<C, M, R>

Constructs a fluent builder for the GetIntentVersions operation.

See GetIntentVersions for more information about the operation and its arguments.

pub fn get_migration(&self) -> GetMigration<C, M, R>

Constructs a fluent builder for the GetMigration operation.

See GetMigration for more information about the operation and its arguments.

pub fn get_migrations(&self) -> GetMigrations<C, M, R>

Constructs a fluent builder for the GetMigrations operation.

See GetMigrations for more information about the operation and its arguments.

pub fn get_slot_type(&self) -> GetSlotType<C, M, R>

Constructs a fluent builder for the GetSlotType operation.

See GetSlotType for more information about the operation and its arguments.

pub fn get_slot_types(&self) -> GetSlotTypes<C, M, R>

Constructs a fluent builder for the GetSlotTypes operation.

See GetSlotTypes for more information about the operation and its arguments.

pub fn get_slot_type_versions(&self) -> GetSlotTypeVersions<C, M, R>

Constructs a fluent builder for the GetSlotTypeVersions operation.

See GetSlotTypeVersions for more information about the operation and its arguments.

pub fn get_utterances_view(&self) -> GetUtterancesView<C, M, R>

Constructs a fluent builder for the GetUtterancesView operation.

See GetUtterancesView for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn put_bot(&self) -> PutBot<C, M, R>

Constructs a fluent builder for the PutBot operation.

See PutBot for more information about the operation and its arguments.

pub fn put_bot_alias(&self) -> PutBotAlias<C, M, R>

Constructs a fluent builder for the PutBotAlias operation.

See PutBotAlias for more information about the operation and its arguments.

pub fn put_intent(&self) -> PutIntent<C, M, R>

Constructs a fluent builder for the PutIntent operation.

See PutIntent for more information about the operation and its arguments.

pub fn put_slot_type(&self) -> PutSlotType<C, M, R>

Constructs a fluent builder for the PutSlotType operation.

See PutSlotType for more information about the operation and its arguments.

pub fn start_import(&self) -> StartImport<C, M, R>

Constructs a fluent builder for the StartImport operation.

See StartImport for more information about the operation and its arguments.

pub fn start_migration(&self) -> StartMigration<C, M, R>

Constructs a fluent builder for the StartMigration operation.

See StartMigration for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.