Module aws_sdk_lexmodelbuilding::error

Modules

access_denied_exception

See AccessDeniedException

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

precondition_failed_exception

See PreconditionFailedException

resource_in_use_exception

See ResourceInUseException

Structs

AccessDeniedException

Your IAM user or role does not have permission to call the Amazon Lex V2 APIs

BadRequestException

The request is not well formed. For example, a value is invalid or

ConflictException

There was a conflict processing the request. Try your request

CreateBotVersionError

CreateIntentVersionError

CreateSlotTypeVersionError

DeleteBotAliasError

DeleteBotChannelAssociationError

DeleteBotError

DeleteBotVersionError

DeleteIntentError

DeleteIntentVersionError

DeleteSlotTypeError

DeleteSlotTypeVersionError

DeleteUtterancesError

GetBotAliasError

GetBotAliasesError

GetBotChannelAssociationError

GetBotChannelAssociationsError

GetBotError

GetBotVersionsError

GetBotsError

GetBuiltinIntentError

GetBuiltinIntentsError

GetBuiltinSlotTypesError

GetExportError

GetImportError

GetIntentError

GetIntentVersionsError

GetIntentsError

GetMigrationError

GetMigrationsError

GetSlotTypeError

GetSlotTypeVersionsError

GetSlotTypesError

GetUtterancesViewError

InternalFailureException

An internal Amazon Lex error occurred. Try your request again.

LimitExceededException

The request exceeded a limit. Try your request again.

ListTagsForResourceError

NotFoundException

The resource specified in the request was not found. Check the

PreconditionFailedException

The checksum of the resource that you are trying to change does

PutBotAliasError

PutBotError

PutIntentError

PutSlotTypeError

ResourceInUseException

The resource that you are attempting to delete is referred to by

StartImportError

StartMigrationError

TagResourceError

UntagResourceError

Enums

CreateBotVersionErrorKind

CreateIntentVersionErrorKind

CreateSlotTypeVersionErrorKind

DeleteBotAliasErrorKind

DeleteBotChannelAssociationErrorKind

DeleteBotErrorKind

DeleteBotVersionErrorKind

DeleteIntentErrorKind

DeleteIntentVersionErrorKind

DeleteSlotTypeErrorKind

DeleteSlotTypeVersionErrorKind

DeleteUtterancesErrorKind

GetBotAliasErrorKind

GetBotAliasesErrorKind

GetBotChannelAssociationErrorKind

GetBotChannelAssociationsErrorKind

GetBotErrorKind

GetBotVersionsErrorKind

GetBotsErrorKind

GetBuiltinIntentErrorKind

GetBuiltinIntentsErrorKind

GetBuiltinSlotTypesErrorKind

GetExportErrorKind

GetImportErrorKind

GetIntentErrorKind

GetIntentVersionsErrorKind

GetIntentsErrorKind

GetMigrationErrorKind

GetMigrationsErrorKind

GetSlotTypeErrorKind

GetSlotTypeVersionsErrorKind

GetSlotTypesErrorKind

GetUtterancesViewErrorKind

ListTagsForResourceErrorKind

PutBotAliasErrorKind

PutBotErrorKind

PutIntentErrorKind

PutSlotTypeErrorKind

StartImportErrorKind

StartMigrationErrorKind

TagResourceErrorKind

UntagResourceErrorKind