Struct aws_sdk_lexmodelbuilding::input::PutIntentInput

#[non_exhaustive]
pub struct PutIntentInput {

    pub name: Option<String>,
    pub description: Option<String>,
    pub slots: Option<Vec<Slot>>,
    pub sample_utterances: Option<Vec<String>>,
    pub confirmation_prompt: Option<Prompt>,
    pub rejection_statement: Option<Statement>,
    pub follow_up_prompt: Option<FollowUpPrompt>,
    pub conclusion_statement: Option<Statement>,
    pub dialog_code_hook: Option<CodeHook>,
    pub fulfillment_activity: Option<FulfillmentActivity>,
    pub parent_intent_signature: Option<String>,
    pub checksum: Option<String>,
    pub create_version: Option<bool>,
    pub kendra_configuration: Option<KendraConfiguration>,
    pub input_contexts: Option<Vec<InputContext>>,
    pub output_contexts: Option<Vec<OutputContext>>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

description: Option<String>

A description of the intent.

slots: Option<Vec<Slot>>

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

sample_utterances: Option<Vec<String>>

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

confirmation_prompt: Option<Prompt>

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

rejection_statement: Option<Statement>

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

follow_up_prompt: Option<FollowUpPrompt>

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

conclusion_statement: Option<Statement>

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

dialog_code_hook: Option<CodeHook>

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

fulfillment_activity: Option<FulfillmentActivity>

Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

parent_intent_signature: Option<String>

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

checksum: Option<String>

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

create_version: Option<bool>

When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

kendra_configuration: Option<KendraConfiguration>

Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

input_contexts: Option<Vec<InputContext>>

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

output_contexts: Option<Vec<OutputContext>>

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

Implementations

impl PutIntentInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<PutIntent, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<PutIntent>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture PutIntentInput