Module aws_sdk_lexmodelbuilding::model

Modules

bot_alias_metadata

See BotAliasMetadata

bot_channel_association

See BotChannelAssociation

bot_metadata

See BotMetadata

builtin_intent_metadata

See BuiltinIntentMetadata

builtin_intent_slot

See BuiltinIntentSlot

builtin_slot_type_metadata

See BuiltinSlotTypeMetadata

code_hook

See CodeHook

conversation_logs_request

See ConversationLogsRequest

conversation_logs_response

See ConversationLogsResponse

enumeration_value

See EnumerationValue

follow_up_prompt

See FollowUpPrompt

fulfillment_activity

See FulfillmentActivity

input_context

See InputContext

intent

See Intent

intent_metadata

See IntentMetadata

kendra_configuration

See KendraConfiguration

log_settings_request

See LogSettingsRequest

log_settings_response

See LogSettingsResponse

message

See Message

migration_alert

See MigrationAlert

migration_summary

See MigrationSummary

output_context

See OutputContext

prompt

See Prompt

resource_reference

See ResourceReference

slot

See Slot

slot_default_value

See SlotDefaultValue

slot_default_value_spec

See SlotDefaultValueSpec

slot_type_configuration

See SlotTypeConfiguration

slot_type_metadata

See SlotTypeMetadata

slot_type_regex_configuration

See SlotTypeRegexConfiguration

statement

See Statement

tag

See Tag

utterance_data

See UtteranceData

utterance_list

See UtteranceList

Structs

BotAliasMetadata

Provides information about a bot alias.

BotChannelAssociation

Represents an association between an Amazon Lex bot and an external

BotMetadata

Provides information about a bot. .

BuiltinIntentMetadata

Provides metadata for a built-in intent.

BuiltinIntentSlot

Provides information about a slot used in a built-in

BuiltinSlotTypeMetadata

Provides information about a built in slot type.

CodeHook

Specifies a Lambda function that verifies requests to a bot or

ConversationLogsRequest

Provides the settings needed for conversation logs.

ConversationLogsResponse

Contains information about conversation log settings.

EnumerationValue

Each slot type can have a set of values. Each enumeration value

FollowUpPrompt

A prompt for additional activity after an intent is fulfilled. For

FulfillmentActivity

Describes how the intent is fulfilled after the user provides all

InputContext

The name of a context that must be active for an intent to be selected

Intent

Identifies the specific version of an intent.

IntentMetadata

Provides information about an intent.

KendraConfiguration

Provides configuration information for the AMAZON.KendraSearchIntent

LogSettingsRequest

Settings used to configure delivery mode and destination for

LogSettingsResponse

The settings for conversation logs.

Message

The message object that provides the message text and its

MigrationAlert

Provides information about alerts and warnings that Amazon Lex sends during

MigrationSummary

Provides information about migrating a bot from Amazon Lex V1 to Amazon Lex V2.

OutputContext

The specification of an output context that is set when an intent is

Prompt

Obtains information from the user. To define a prompt, provide one

ResourceReference

Describes the resource that refers to the resource that you are

Slot

Identifies the version of a specific slot.

SlotDefaultValue

A default value for a slot.

SlotDefaultValueSpec

Contains the default values for a slot. Default values are used when

SlotTypeConfiguration

Provides configuration information for a slot type.

SlotTypeMetadata

Provides information about a slot type..

SlotTypeRegexConfiguration

Provides a regular expression used to validate the value of a

Statement

A collection of messages that convey information to the user. At

Tag

A list of key/value pairs that identify a bot, bot alias, or bot

UtteranceData

Provides information about a single utterance that was made to your

UtteranceList

Provides a list of utterances that have been made to a specific

Enums

ChannelStatus

ChannelType

ContentType

Destination

ExportStatus

ExportType

FulfillmentActivityType

ImportStatus

Locale

LogType

MergeStrategy

MigrationAlertType

MigrationSortAttribute

MigrationStatus

MigrationStrategy

ObfuscationSetting

ProcessBehavior

ReferenceType

ResourceType

SlotConstraint

SlotValueSelectionStrategy

SortOrder

Status

StatusType