Module aws_sdk_lexmodelbuilding::operation

Structs

CreateBotVersion

Creates a new version of the bot based on the $LATEST

CreateIntentVersion

Creates a new version of an intent based on the

CreateSlotTypeVersion

Creates a new version of a slot type based on the

DeleteBot

Deletes all versions of the bot, including the $LATEST

DeleteBotAlias

Deletes an alias for the specified bot.

DeleteBotChannelAssociation

Deletes the association between an Amazon Lex bot and a messaging

DeleteBotVersion

Deletes a specific version of a bot. To delete all versions of a

DeleteIntent

Deletes all versions of the intent, including the

DeleteIntentVersion

Deletes a specific version of an intent. To delete all versions of

DeleteSlotType

Deletes all versions of the slot type, including the

DeleteSlotTypeVersion

Deletes a specific version of a slot type. To delete all versions

DeleteUtterances

Deletes stored utterances.

GetBot

Returns metadata information for a specific bot. You must provide

GetBotAlias

Returns information about an Amazon Lex bot alias. For more information

GetBotAliases

Returns a list of aliases for a specified Amazon Lex bot.

GetBotChannelAssociation

Returns information about the association between an Amazon Lex bot and

GetBotChannelAssociations

Returns a list of all of the channels associated with the

GetBotVersions

Gets information about all of the versions of a bot.

GetBots

Returns bot information as follows:

GetBuiltinIntent

Returns information about a built-in intent.

GetBuiltinIntents

Gets a list of built-in intents that meet the specified

GetBuiltinSlotTypes

Gets a list of built-in slot types that meet the specified

GetExport

Exports the contents of a Amazon Lex resource in a specified format.

GetImport

Gets information about an import job started with the

GetIntent

Returns information about an intent. In addition to the intent

GetIntentVersions

Gets information about all of the versions of an intent.

GetIntents

Returns intent information as follows:

GetMigration

Provides details about an ongoing or complete migration from an

GetMigrations

Gets a list of migrations between Amazon Lex V1 and Amazon Lex V2.

GetSlotType

Returns information about a specific version of a slot type. In

GetSlotTypeVersions

Gets information about all versions of a slot type.

GetSlotTypes

Returns slot type information as follows:

GetUtterancesView

Use the GetUtterancesView operation to get information

ListTagsForResource

Gets a list of tags associated with the specified resource. Only bots,

PutBot

Creates an Amazon Lex conversational bot or replaces an existing bot.

PutBotAlias

Creates an alias for the specified version of the bot or replaces

PutIntent

Creates an intent or replaces an existing intent.

PutSlotType

Creates a custom slot type or replaces an existing custom slot

StartImport

Starts a job to import a resource to Amazon Lex.

StartMigration

Starts migrating a bot from Amazon Lex V1 to Amazon Lex V2. Migrate your bot when

TagResource

Adds the specified tags to the specified resource. If a tag key

UntagResource

Removes tags from a bot, bot alias or bot channel.