Struct aws_sdk_lexmodelbuilding::client::fluent_builders::PutIntent

pub struct PutIntent<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to PutIntent.

Creates an intent or replaces an existing intent.

To define the interaction between the user and your bot, you use one or more intents. For a pizza ordering bot, for example, you would create an OrderPizza intent.

To create an intent or replace an existing intent, you must provide the following:

Intent name. For example, OrderPizza.

Sample utterances. For example, "Can I order a pizza, please." and "I want to order a pizza."

Information to be gathered. You specify slot types for the information that your bot will request from the user. You can specify standard slot types, such as a date or a time, or custom slot types such as the size and crust of a pizza.

How the intent will be fulfilled. You can provide a Lambda function or configure the intent to return the intent information to the client application. If you use a Lambda function, when all of the intent information is available, Amazon Lex invokes your Lambda function. If you configure your intent to return the intent information to the client application.

You can specify other optional information in the request, such as:

A confirmation prompt to ask the user to confirm an intent. For example, "Shall I order your pizza?"

A conclusion statement to send to the user after the intent has been fulfilled. For example, "I placed your pizza order."

A follow-up prompt that asks the user for additional activity. For example, asking "Do you want to order a drink with your pizza?"

If you specify an existing intent name to update the intent, Amazon Lex replaces the values in the $LATEST version of the intent with the values in the request. Amazon Lex removes fields that you don't provide in the request. If you don't specify the required fields, Amazon Lex throws an exception. When you update the $LATEST version of an intent, the status field of any bot that uses the $LATEST version of the intent is set to NOT_BUILT.

For more information, see how-it-works.

This operation requires permissions for the lex:PutIntent action.

Implementations

impl<C, M, R> PutIntent<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<PutIntentOutput, SdkError<PutIntentError>> where
R::Policy: SmithyRetryPolicy<PutIntentInputOperationOutputAlias, PutIntentOutput, PutIntentError, PutIntentInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn name(self, inp: impl Into<String>) -> Self

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

pub fn set_name(self, input: Option<String>) -> Self

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

pub fn description(self, inp: impl Into<String>) -> Self

A description of the intent.

pub fn set_description(self, input: Option<String>) -> Self

A description of the intent.

pub fn slots(self, inp: impl Into<Slot>) -> Self

Appends an item to slots.

To override the contents of this collection use set_slots.

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

pub fn set_slots(self, input: Option<Vec<Slot>>) -> Self

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

pub fn sample_utterances(self, inp: impl Into<String>) -> Self

Appends an item to sampleUtterances.

To override the contents of this collection use set_sample_utterances.

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

pub fn set_sample_utterances(self, input: Option<Vec<String>>) -> Self

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

pub fn confirmation_prompt(self, inp: Prompt) -> Self

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

pub fn set_confirmation_prompt(self, input: Option<Prompt>) -> Self

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

pub fn rejection_statement(self, inp: Statement) -> Self

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

pub fn set_rejection_statement(self, input: Option<Statement>) -> Self

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

pub fn follow_up_prompt(self, inp: FollowUpPrompt) -> Self

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

pub fn set_follow_up_prompt(self, input: Option<FollowUpPrompt>) -> Self

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

pub fn conclusion_statement(self, inp: Statement) -> Self

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

pub fn set_conclusion_statement(self, input: Option<Statement>) -> Self

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

pub fn dialog_code_hook(self, inp: CodeHook) -> Self

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

pub fn set_dialog_code_hook(self, input: Option<CodeHook>) -> Self

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

pub fn fulfillment_activity(self, inp: FulfillmentActivity) -> Self

Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

pub fn set_fulfillment_activity(
self,
input: Option<FulfillmentActivity>
) -> Self

Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

pub fn parent_intent_signature(self, inp: impl Into<String>) -> Self

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

pub fn set_parent_intent_signature(self, input: Option<String>) -> Self

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

pub fn checksum(self, inp: impl Into<String>) -> Self

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

pub fn set_checksum(self, input: Option<String>) -> Self

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

pub fn create_version(self, inp: bool) -> Self

When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

pub fn set_create_version(self, input: Option<bool>) -> Self

When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

pub fn kendra_configuration(self, inp: KendraConfiguration) -> Self

Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

pub fn set_kendra_configuration(
self,
input: Option<KendraConfiguration>
) -> Self

Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

pub fn input_contexts(self, inp: impl Into<InputContext>) -> Self

Appends an item to inputContexts.

To override the contents of this collection use set_input_contexts.

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

pub fn set_input_contexts(self, input: Option<Vec<InputContext>>) -> Self

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

pub fn output_contexts(self, inp: impl Into<OutputContext>) -> Self

Appends an item to outputContexts.

To override the contents of this collection use set_output_contexts.

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

pub fn set_output_contexts(self, input: Option<Vec<OutputContext>>) -> Self

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.